Add files via upload
So this is what we pulled off the video. It allows for local messaging, so you can send yourself a message from the same machine. We attempted to send a message my computer to his but were unable to transmit the message successfully. There was difficulty in trying to get it to connect to our machine. We might need to look into port forwarding for this next bit.